requirements.txt
pandas
scikit-learn
joblib
requests
python-whois
customtkinter
python-dotenv
validators
cefpython3